refactor: 매퍼 파일 이동
