+++ presentation/src/main/java/com/paykidscompose/presentation/mapper/my/MyInfoUIModelMapper.kt
@@ -1,0 +1,19 @@
+package com.paykidscompose.presentation.mapper.my
+
+import com.paykidscompose.common.mapper.ModelMapper
+import com.paykidscompose.common.model.user.UserModel
+import com.paykidscompose.presentation.model.MyInfoUIModel
+
+object MyInfoUIModelMapper : ModelMapper<UserModel, MyInfoUIModel> {
+    override fun mapToModel(layerModel: MyInfoUIModel): UserModel {
+        throw UnsupportedOperationException("MyInfoUIModel -> UserModel 변환은 지원하지 않습니다.")
+    }
+
+    override fun mapToLayerModel(model: UserModel): MyInfoUIModel {
+        return MyInfoUIModel(
+            nickname = model.nickname,
+            email = model.email,
+            image = model.profileImageUrl
+        )
+    }
+}
