+++ presentation/src/main/java/com/paykidscompose/presentation/screens/mypage/info/MyInfoViewModel.kt
@@ -9,8 +9,8 @@ import com.paykidscompose.common.usecase.user.GetUserUseCase
 import com.paykidscompose.common.usecase.user.ReplaceNicknameUseCase
 import com.paykidscompose.presentation.base.UIEvent
 import com.paykidscompose.presentation.base.UIState
+import com.paykidscompose.presentation.mapper.my.MyInfoUIModelMapper
 import com.paykidscompose.presentation.model.MyInfoUIModel
-import com.paykidscompose.presentation.model.MyInfoUIModelMapper
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asSharedFlow
